Revert kubebuilder minimal length tag and go back to custom validation
