+++ config/crd/bases/dynatrace.com_edgeconnects.yaml
@@ -963,7 +963,6 @@ spec:
                 default: dynatrace-edgeconnect
                 description: ServiceAccountName that allows EdgeConnect to access
                   the Kubernetes API
-                minLength: 1
                 type: string
               tolerations:
                 description: Sets tolerations for the EdgeConnect pods
