+++ config/helm/chart/default/templates/Common/crd/dynatrace-operator-crd.yaml
@@ -8312,7 +8312,6 @@ spec:
                 default: dynatrace-edgeconnect
                 description: ServiceAccountName that allows EdgeConnect to access
                   the Kubernetes API
-                minLength: 1
                 type: string
               tolerations:
                 description: Sets tolerations for the EdgeConnect pods
