+++ pkg/api/v1alpha2/edgeconnect/validation/api_server_test.go
@@ -4,7 +4,6 @@ import (
 	"testing"
 
 	"github.com/Dynatrace/dynatrace-operator/pkg/api/v1alpha2/edgeconnect"
-	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
@@ -12,15 +11,6 @@ const (
 	testName      = "test-name"
 	testNamespace = "test-namespace"
 )
-
-func prepareTestServiceAccount(name string, namespace string) *corev1.ServiceAccount {
-	return &corev1.ServiceAccount{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      name,
-			Namespace: namespace,
-		},
-	}
-}
 
 func TestApiServer(t *testing.T) {
 	t.Run(`happy apiServer`, func(t *testing.T) {
