+++ pkg/api/v1alpha2/edgeconnect/validation/validation.go
@@ -23,6 +23,7 @@ var validatorErrorFuncs = []validatorFunc{
 	isInvalidApiServer,
 	nameTooLong,
 	checkHostPatternsValue,
+	isInvalidServiceName,
 	automationRequiresProvisionerValidation,
 }
